se diseño el cardGameCheckPopint
