+++ app/src/main/java/com/example/comusenias/presentation/ui/theme/Size.kt
@@ -6,12 +6,14 @@ const val SIZE2: Int = 1
 const val size3: Int = 3
 const val size5: Int = 5
 const val size6: Int = 6
+const val size7: Int = 7
 const val size10: Int = 10
 const val SIZE12: Int = 12
 const val size14: Int = 14
 const val size15: Int = 15
 const val SIZE16: Int = 16
 const val size17: Int = 17
+const val size18: Int = 18
 const val size19: Int = 19
 const val size20: Int = 20
 const val SIZE22: Int = 22
